+++ R/request-make.R
@@ -1,9 +1,15 @@
 #' Make a Google API request
 #'
+#' @description
 #' Intended primarily for internal use in client packages that provide
-#' high-level wrappers for users. `request_make()` does very little: calls an
-#' HTTP method, only adding a user agent. Typically the input is created with
-#' [request_build()] and the output is processed with [response_process()].
+#' high-level wrappers for users. `request_make()` does relatively little:
+#' * Calls an HTTP method.
+#' * Adds a user agent.
+#' * Enforces `"json"` as the default for `encode`. This differs from httr's
+#'   default behaviour, but aligns better with Google APIs.
+#'
+#' Typically the input is created with [request_build()] and the output is
+#' processed with [response_process()].
 #'
 #' @param x List. Holds the components for an HTTP request, presumably created
 #'   with [request_develop()] or [request_build()]. Must contain a `method` and
@@ -12,7 +18,10 @@
 #'   in doubt, a client package should have an internal function that extends
 #'   `gargle_user_agent()` by prepending its return value with the client
 #'   package's name and version.
-#' @param ... Optional arguments passed through to the HTTP method.
+#' @inheritParams httr::POST
+#' @param ... Optional arguments passed through to the HTTP method. Currently
+#'   neither gargle nor httr checks that all are used, so be aware that unused
+#'   arguments may be silently ignored.
 #'
 #' @return Object of class `response` from [httr].
 #' @export
@@ -26,7 +35,10 @@
 #' )
 #' gargle::request_make(req)
 #' }
-request_make <- function(x, ..., user_agent = gargle_user_agent()) {
+request_make <- function(x,
+                         ...,
+                         encode = "json",
+                         user_agent = gargle_user_agent()) {
   stopifnot(is.character(x$method))
   method <- switch(
     x$method,
@@ -41,6 +53,7 @@ request_make <- function(x, ..., user_agent = gargle_user_agent()) {
     url = x$url,
     body = x$body,
     x$token,
+    encode = encode,
     user_agent,
     ...
   )